refactor: move control and merge package to util
