+++ internal/util/merge/slice_test.go
@@ -3,7 +3,7 @@ package merge_test
 import (
 	"testing"
 
-	"github.com/hetznercloud/terraform-provider-hcloud/internal/merge"
+	"github.com/hetznercloud/terraform-provider-hcloud/internal/util/merge"
 	"github.com/stretchr/testify/assert"
 )
